macOS Mojave button text fix

+++ build_macOS.sh
@@ -4,6 +4,31 @@
 # It is here because I am lazy
 # ---------------------------------------------------------------------
 
+# Call getopt to validate the provided input. 
+VERBOSE=false
+MAKE_DISK=false
+KEEP_VENV=false
+while getopts "hvde" OPTION; do
+	case "$OPTION" in
+		h)  echo "Options:"
+			echo "\t-h Print help (this)"
+			echo "\t-v Verbose output"
+			echo "\t-e Keep Python virtual environment (don't delete)"
+			echo "\t-d Make disk image (.dmg)"
+			exit 0
+			;;
+		v) 	VERBOSE=true
+			;;
+		d) 	MAKE_DISK=true
+			;;
+		e)  KEEP_VENV=true
+			;;
+		*)  echo "Incorrect option provided"
+			exit 1
+			;;
+    esac
+done
+
 echo "Validate environment..."
 
 # Get version from main source file.
@@ -36,7 +61,7 @@ fi
 
 # Clean up any previous build work
 echo "Remove old builds..."
-rm -rf ./build ./dist *.pyc
+rm -rf ./build ./dist *.pyc ./__pycache__
 
 # Set up and activate virtual environment for dependencies
 echo "Setup Python Virtual Environment..."
@@ -62,16 +87,23 @@ echo "Copy support files..."
 cp k40_whisperer_test.svg Change_Log.txt gpl-3.0.txt README_MacOS.md dist
 
 # Clean up the build directory when we are done.
-echo "Clean up and deactivate Python virtual environment..."
+echo "Clean up build artifacts..."
 rm -rf build
 
 # Remove virtual environment
-deactivate
-rm -rf python_venv
+if [ "$KEEP_VENV" = false ]
+then
+	echo "Remove Python virtual environment..."
+	deactivate
+	rm -rf python_venv
+fi
 
 # Buid a new disk image
-echo "Build macOS Disk Image..."
-rm ./K40-Whisperer-${VERSION}.dmg
-hdiutil create -fs HFS+ -volname K40-Whisperer-${VERSION} -srcfolder ./dist ./K40-Whisperer-${VERSION}.dmg
+if [ "$MAKE_DISK" = true ]
+then
+	echo "Build macOS Disk Image..."
+	rm ./K40-Whisperer-${VERSION}.dmg
+	hdiutil create -fs HFS+ -volname K40-Whisperer-${VERSION} -srcfolder ./dist ./K40-Whisperer-${VERSION}.dmg
+fi
 
 echo "Done."
